+++ const_format/src/lib.rs
@@ -205,6 +205,7 @@
 //! So `#[doc = "foobar"]` cannot be replaced with `#[doc = concatcp!("foo", "bar") ]`.
 //!
 //! <span id="integer-args"></span>
+//!
 //! ### Integer arguments
 //!
 //! Integer arguments must have a type inferrable from context.
@@ -222,6 +223,18 @@
 //! Example of what does not compile:
 //! ```compile_fail
 //! assert_eq!(const_format::concatcp!(1 + 1, 2 + 1), "23");
+//! ```
+//!
+//! # Renaming crate
+//!
+//! All function-like macros from `const_format` can be used when the crate is renamed.
+//!
+//! The [`ConstDebug`] derive macro has the `#[cdeb(crate = "foo::bar")]` attribute to
+//! tell it where to find the `const_format` crate.
+//!
+//! Example of renaming the `const_format` crate in the Cargo.toml file:
+//! ```toml
+//! cfmt = {version = "0.*", package = "const_format"}
 //! ```
 //!
 //! # Cargo features