+++ const_format/tests/fmt_tests/formatted_writing.rs
@@ -1,4 +1,4 @@
-use const_format::{
+use cfmt_a::{
     __for_range,
     fmt::{ComputeStrLength, Error, Formatter, FormattingFlags, StrWriter},
     try_,