+++ const_format/tests/misc_tests/type_kind_coercion_macro_tests.rs
@@ -1,5 +1,5 @@
-use const_format::fmt::{Error, Formatter, FormattingFlags, StrWriter};
-use const_format::{coerce_to_fmt, impl_fmt};
+use cfmt_b::fmt::{Error, Formatter, FormattingFlags, StrWriter};
+use cfmt_b::{coerce_to_fmt, impl_fmt};
 
 #[test]
 fn coercion() {