+++ const_format/tests/misc_tests_modules.rs
@@ -1,7 +1,15 @@
 #![cfg_attr(feature = "fmt", feature(const_mut_refs))]
 
+extern crate const_format as cfmt_b;
+extern crate self as const_format;
+
+// Making sure that `const_format` points at this test crate.
+pub const NOT_CF: usize = 13;
+pub const _ASSERT_NOT_CF: [(); 13] = [(); const_format::NOT_CF];
+
 mod misc_tests {
     #[cfg(feature = "assert")]
+    #[cfg(not(feature = "only_new_tests"))]
     mod asserts;
 
     #[cfg(feature = "fmt")]
@@ -9,12 +17,15 @@ mod misc_tests {
     mod call_debug_fmt_macro;
 
     #[cfg(feature = "fmt")]
+    #[cfg(not(feature = "only_new_tests"))]
     mod concatc_macro_tests;
 
     #[cfg(feature = "derive")]
+    #[cfg(not(feature = "only_new_tests"))]
     mod derive_tests;
 
     #[cfg(feature = "assert")]
+    #[cfg(not(feature = "only_new_tests"))]
     mod equality_tests;
 
     #[cfg(not(feature = "only_new_tests"))]
@@ -24,7 +35,8 @@ mod misc_tests {
     #[cfg(not(feature = "only_new_tests"))]
     mod impl_fmt_macro_tests;
 
-    // #[cfg(not(feature = "only_new_tests"))]
+    #[cfg(feature = "assert")]
+    #[cfg(not(feature = "only_new_tests"))]
     mod shared_cp_macro_tests;
 
     #[cfg(feature = "fmt")]
@@ -32,6 +44,6 @@ mod misc_tests {
     mod type_kind_coercion_macro_tests;
 
     #[cfg(feature = "fmt")]
-    #[cfg(not(feature = "only_new_tests"))]
+    //#[cfg(not(feature = "only_new_tests"))]
     mod writec_macro;
 }