+++ print_errors/src/lib.rs
@@ -3,10 +3,10 @@
 
 mod formatc_macros;
 
-const _: &str = const_format::concatcp!(0, 1, ());
+const _: &str = cfmt::concatcp!(0, 1, ());
 
 #[cfg(feature = "nightly")]
-const _: &str = const_format::concatc!(0, 1, ());
+const _: &str = cfmt::concatc!(0, 1, ());
 
 #[cfg(feature = "nightly")]
 mod using_assertc_macros;