+++ print_errors/src/using_writec_macro.rs
@@ -1,6 +1,6 @@
-use const_format::{writec, StrWriter};
+use cfmt::{writec, StrWriter};
 
-fn using_writec(writer: &mut StrWriter) -> const_format::Result {
+fn using_writec(writer: &mut StrWriter) -> cfmt::Result {
     // Trying to write to a non-writer
     writec!((), "")?;
 
@@ -10,6 +10,12 @@ fn using_writec(writer: &mut StrWriter) -> const_format::Result {
 
     // trying to write an uninferred integer type
     writec!(writer, "{}", 0)?;
+
+    let a = 0;
+    writec!(writer, "{}", a)?;
+    writec!(writer, "{b}", b = a)?;
+    writec!(writer, "{a}")?;
+    writec!(writer, "{a:?}")?;
 
     Ok(())
 }